requirements.txt
streamlit
pandas
numpy
plotly
google-cloud-bigquery
requests